+++ .rustfmt.toml
@@ -1,3 +1,5 @@
 hard_tabs = true
 max_width = 80
 edition = "2018"
+imports_granularity = "Crate"
+group_imports = "StdExternalCrate"
