+++ src/github/client/file.rs
@@ -1,8 +1,7 @@
-use crate::{github::*, Result};
+use super::GithubClient;
+use crate::{github::*, types::Result};
 
-use super::GithubBot;
-
-impl GithubBot {
+impl GithubClient {
 	pub async fn contents(
 		&self,
 		owner: &str,
