+++ src/github/client/pull_request.rs
@@ -1,21 +1,20 @@
+use super::GithubClient;
 use crate::{
 	companion::CompanionReferenceTrailItem,
+	config::MainConfig,
 	error::Error,
 	github::*,
-	github_bot::MainConfig,
-	webhook::{Dependency, MergeRequest},
-	Result,
+	merge_request::{MergeRequest, MergeRequestDependency},
+	types::Result,
 };
 
-use super::GithubBot;
-
-impl GithubBot {
+impl GithubClient {
 	pub async fn pull_request(
 		&self,
 		owner: &str,
 		repo: &str,
 		number: i64,
-	) -> Result<PullRequest> {
+	) -> Result<GithubPullRequest> {
 		self.client
 			.get(format!(
 				"{}/repos/{}/{}/pulls/{}",
@@ -29,7 +28,7 @@ impl GithubBot {
 		owner: &str,
 		repo: &str,
 		head: &str,
-	) -> Result<Option<PullRequest>> {
+	) -> Result<Option<GithubPullRequest>> {
 		self.client
 			.get_all(format!(
 				"{}/repos/{}/{}/pulls?head={}",
@@ -60,7 +59,7 @@ impl GithubBot {
 	pub async fn resolve_pr_dependents(
 		&self,
 		config: &MainConfig,
-		pr: &PullRequest,
+		pr: &GithubPullRequest,
 		requested_by: &str,
 		companion_reference_trail: &[CompanionReferenceTrailItem],
 	) -> Result<Option<Vec<MergeRequest>>, Error> {
@@ -70,7 +69,7 @@ impl GithubBot {
 				None => return Ok(None),
 			};
 
-		let parent_dependency = Dependency {
+		let parent_dependency = MergeRequestDependency {
 			sha: (&pr.head.sha).into(),
 			owner: (&pr.base.repo.owner.login).into(),
 			repo: (&pr.base.repo.name).into(),
@@ -80,19 +79,17 @@ impl GithubBot {
 		};
 		let dependents =
 			// If there's only one companion, then it can't possibly depend on another companion
-			if let [(comp_html_url, comp_owner, comp_repo, comp_number)] =
-				&*companions
-			{
+			if let [comp] = &*companions {
 				let comp_pr = self
-					.pull_request(comp_owner, comp_repo, *comp_number)
+					.pull_request(&comp.owner, &comp.repo, comp.number)
 					.await?;
 				vec![MergeRequest {
 					was_updated: false,
 					sha: comp_pr.head.sha,
-					owner: comp_owner.into(),
-					repo: comp_repo.into(),
-					number: *comp_number,
-					html_url: comp_html_url.into(),
+					owner: comp_pr.base.repo.owner.login,
+					repo: comp_pr.base.repo.name,
+					number: comp_pr.number,
+					html_url: comp_pr.html_url,
 					requested_by: requested_by.into(),
 					dependencies: Some(vec![parent_dependency]),
 				}]
@@ -100,16 +97,19 @@ impl GithubBot {
 				let base_dependencies = vec![parent_dependency];
 
 				let mut dependents = vec![];
-				for (comp_html_url, comp_owner, comp_repo, comp_number) in &companions {
+				for comp in &companions {
 					// Prevent duplicate dependencies in case of error in user input
-					if comp_repo == &pr.base.repo.owner.login {
+					if comp.repo == pr.base.repo.owner.login {
 						continue;
 					}
 
 					// Fetch the companion's lockfile in order to detect its dependencies
 					let comp_pr = self
-						.pull_request(comp_owner, comp_repo, *comp_number)
+						.pull_request(&comp.owner, &comp.repo, comp.number)
 						.await?;
+					let comp_owner = &comp_pr.base.repo.owner.login;
+					let comp_repo = &comp_pr.base.repo.name;
+
 					let comp_lockfile = {
 						let lockfile_content = self
 							.contents(
@@ -124,16 +124,16 @@ impl GithubBot {
 						)
 						.map_err(|err| Error::Message {
 							msg: format!(
-								"Failed to decode the API content for the lockfile of {}/{}/pull/{}: {:?}",
-								comp_owner, comp_repo, comp_number, err
+								"Failed to decode the API content for the lockfile of {}: {:?}",
+								&comp_pr.html_url, err
 							),
 						})?;
 						let txt = String::from_utf8_lossy(&txt_encoded);
 						txt.parse::<cargo_lock::Lockfile>().map_err(|err| {
 							Error::Message {
 								msg: format!(
-								"Failed to parse lockfile of {}/{}/pull/{}: {:?}",
-								comp_owner, comp_repo, comp_number, err
+								"Failed to parse lockfile of {}: {:?}",
+								&comp_pr.html_url, err
 							),
 							}
 						})?
@@ -143,22 +143,17 @@ impl GithubBot {
 
 					// Go through all the other companions to check if any of them is a dependency
 					// of this companion
-					'to_next_other_companion: for (
-						other_comp_html_url,
-						other_comp_owner,
-						other_comp_repo,
-						other_comp_number,
-					) in &companions
+					'to_next_other_companion: for other_comp in &companions
 					{
-						if other_comp_repo == comp_repo ||
+						if &other_comp.repo == comp_repo ||
 							// Prevent duplicate dependencies in case of error in user input
-							other_comp_repo == &pr.base.repo.owner.login {
+							other_comp.repo == pr.base.repo.owner.login {
 							continue;
 						}
 						let other_comp_github_url = format!(
 							"{}/{}/{}{}",
 							config.github_source_prefix,
-							other_comp_owner, other_comp_repo,
+							&other_comp.owner, &other_comp.repo,
 							config.github_source_suffix
 						);
 						for pkg in comp_lockfile.packages.iter() {
@@ -166,17 +161,17 @@ impl GithubBot {
 								if src.url().as_str() == other_comp_github_url {
 									let other_comp_pr = self
 										.pull_request(
-											other_comp_owner,
-											other_comp_repo,
-											*other_comp_number,
+											&other_comp.owner,
+											&other_comp.repo,
+											other_comp.number,
 										)
 										.await?;
-									dependencies.push(Dependency {
-										owner: other_comp_owner.into(),
-										repo: other_comp_repo.into(),
+									dependencies.push(MergeRequestDependency {
+										owner: other_comp_pr.base.repo.owner.login,
+										repo: other_comp_pr.base.repo.name,
 										sha: other_comp_pr.head.sha,
-										number: *other_comp_number,
-										html_url: other_comp_html_url.into(),
+										number: other_comp_pr.number,
+										html_url: other_comp_pr.html_url,
 										is_directly_referenced: false
 									});
 									continue 'to_next_other_companion;
@@ -190,8 +185,8 @@ impl GithubBot {
 						sha: comp_pr.head.sha,
 						owner: comp_owner.into(),
 						repo: comp_repo.into(),
-						number: *comp_number,
-						html_url: comp_html_url.into(),
+						number: comp_pr.number,
+						html_url: comp_pr.html_url,
 						requested_by: requested_by.into(),
 						dependencies: Some(dependencies),
 					})
