+++ src/github/mod.rs
@@ -1,35 +1,39 @@
-use crate::{
-	companion::{parse_all_companions, CompanionReferenceTrailItem},
-	error::*,
-	utils::parse_bot_comment_from_text,
-	PlaceholderDeserializationItem, OWNER_AND_REPO_SEQUENCE, PR_HTML_URL_REGEX,
-};
 use regex::Regex;
 use serde::{Deserialize, Serialize};
 
-pub trait HasIssueDetails {
-	fn get_issue_details(&self) -> Option<IssueDetails>;
+use crate::{
+	bot::parse_bot_comment_from_text,
+	companion::{parse_all_companions, CompanionReferenceTrailItem},
+	error::*,
+	types::PlaceholderDeserializationItem,
+	OWNER_AND_REPO_SEQUENCE, PR_HTML_URL_REGEX,
+};
+
+pub mod client;
+
+pub trait HasPullRequestDetails {
+	fn get_issue_details(&self) -> Option<PullRequestDetails>;
 }
 
 #[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct PullRequest {
+pub struct GithubPullRequest {
 	pub url: String,
 	pub html_url: String,
 	pub number: i64,
-	pub user: Option<User>,
+	pub user: Option<GithubUser>,
 	pub body: Option<String>,
-	pub head: Head,
-	pub base: Base,
+	pub head: GithubPullRequestHead,
+	pub base: GithubPullRequestBase,
 	pub mergeable: Option<bool>,
 	pub merged: bool,
 	pub maintainer_can_modify: bool,
 }
 
-impl PullRequest {
+impl GithubPullRequest {
 	pub fn parse_all_companions(
 		&self,
 		companion_reference_trail: &[CompanionReferenceTrailItem],
-	) -> Option<Vec<IssueDetailsWithRepositoryURL>> {
+	) -> Option<Vec<PullRequestDetailsWithHtmlUrl>> {
 		let mut next_trail =
 			Vec::with_capacity(companion_reference_trail.len() + 1);
 		next_trail.extend_from_slice(companion_reference_trail);
@@ -44,86 +48,26 @@ impl PullRequest {
 }
 
 #[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct Issue {
-	pub number: i64,
-	pub html_url: String,
-	// User might be missing when it has been deleted
-	pub user: Option<User>,
-	pub body: Option<String>,
-	pub pull_request: Option<PlaceholderDeserializationItem>,
-	pub repository: Option<Repository>,
-	pub repository_url: Option<String>,
-}
-
-#[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
 pub struct Contents {
 	pub content: String,
 }
 
 #[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum Event {
-	AddedToProject,
-	RemovedFromProject,
-	#[serde(other)]
-	Unknown,
-}
-
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct IssueEvent {
-	pub project_card: Option<ProjectCard>,
-	pub event: Option<Event>,
-}
-
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct Comment {
+pub struct GithubComment {
 	pub body: String,
 	// User might be missing when it has been deleted
-	pub user: Option<User>,
-}
-
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct ProjectCard {
-	pub id: Option<i64>,
-	pub project_id: Option<i64>,
-	pub project_url: Option<String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct Project {
-	pub id: Option<i64>,
-	pub name: String,
-	pub columns_url: Option<String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub enum ProjectCardContentType {
-	Issue,
-	PullRequest,
-}
-
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct ProjectColumn {
-	pub name: Option<String>,
+	pub user: Option<GithubUser>,
 }
 
 #[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct Head {
-	pub sha: String,
-	pub repo: HeadRepo,
+pub struct GithubPullRequestBase {
 	#[serde(rename = "ref")]
 	pub ref_field: String,
-}
-
-#[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct Base {
-	#[serde(rename = "ref")]
-	pub ref_field: String,
-	pub repo: BaseRepo,
+	pub repo: GithubPullRequestBaseRepository,
 }
 
 #[derive(PartialEq, Debug, Clone, Serialize, Deserialize)]
-pub enum UserType {
+pub enum GithubUserType {
 	User,
 	Bot,
 	#[serde(other)]
@@ -131,37 +75,32 @@ pub enum UserType {
 }
 
 #[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct User {
+pub struct GithubUser {
 	pub login: String,
 	#[serde(rename = "type")]
-	pub type_field: Option<UserType>,
+	pub type_field: Option<GithubUserType>,
 }
 
 #[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct Label {
-	pub name: String,
-}
-
-#[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct Repository {
+pub struct GithubRepository {
 	pub name: String,
 	pub full_name: String,
-	pub owner: User,
+	pub owner: GithubUser,
 	pub html_url: String,
 }
 
 #[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct Status {
+pub struct GithubCommitStatus {
 	pub id: i64,
 	pub context: String,
-	pub state: StatusState,
+	pub state: GithubCommitStatusState,
 	pub description: Option<String>,
 	pub target_url: Option<String>,
 }
 
 #[derive(Debug, Copy, Clone, PartialEq, Serialize, Deserialize)]
 #[serde(rename_all = "snake_case")]
-pub enum StatusState {
+pub enum GithubCommitStatusState {
 	Success,
 	Error,
 	Failure,
@@ -171,13 +110,13 @@ pub enum StatusState {
 
 #[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
 pub struct InstallationRepositories {
-	pub repositories: Vec<Repository>,
+	pub repositories: Vec<GithubRepository>,
 }
 
 #[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
 pub struct Installation {
 	pub id: i64,
-	pub account: User,
+	pub account: GithubUser,
 }
 
 #[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
@@ -203,7 +142,7 @@ pub struct RefObject {
 
 #[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
 #[serde(rename_all = "snake_case")]
-pub enum IssueCommentAction {
+pub enum GithubIssueCommentAction {
 	Created,
 	#[serde(other)]
 	Unknown,
@@ -211,24 +150,32 @@ pub enum IssueCommentAction {
 
 #[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
 pub struct CheckRuns {
-	pub check_runs: Vec<CheckRun>,
+	pub check_runs: Vec<GithubCheckRun>,
 }
 
 #[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct HeadRepo {
+pub struct GithubPullRequestHeadRepository {
 	pub name: String,
-	pub owner: User,
+	pub owner: GithubUser,
 }
 
 #[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct BaseRepo {
+pub struct GithubPullRequestHead {
+	pub sha: String,
+	pub repo: GithubPullRequestHeadRepository,
+	#[serde(rename = "ref")]
+	pub ref_field: String,
+}
+
+#[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
+pub struct GithubPullRequestBaseRepository {
 	pub name: String,
-	pub owner: User,
+	pub owner: GithubUser,
 }
 
 #[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
 #[serde(rename_all = "snake_case")]
-pub enum CheckRunConclusion {
+pub enum GithubCheckRunConclusion {
 	Success,
 	#[serde(other)]
 	Unknown,
@@ -236,54 +183,53 @@ pub enum CheckRunConclusion {
 
 #[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
 #[serde(rename_all = "snake_case")]
-pub enum CheckRunStatus {
+pub enum GithubCheckRunStatus {
 	Completed,
 	#[serde(other)]
 	Unknown,
 }
 
 #[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct CheckRun {
+pub struct GithubCheckRun {
 	pub id: i64,
 	pub name: String,
-	pub status: CheckRunStatus,
-	pub conclusion: Option<CheckRunConclusion>,
+	pub status: GithubCheckRunStatus,
+	pub conclusion: Option<GithubCheckRunConclusion>,
 	pub head_sha: String,
 }
 
 #[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct WebhookIssueComment {
+pub struct GithubWebhookIssueComment {
 	pub number: i64,
 	pub html_url: String,
 	pub repository_url: String,
 	pub pull_request: Option<PlaceholderDeserializationItem>,
 }
-
-impl HasIssueDetails for WebhookIssueComment {
-	fn get_issue_details(&self) -> Option<IssueDetails> {
-		parse_issue_details_from_pr_html_url(&self.html_url)
+impl HasPullRequestDetails for GithubWebhookIssueComment {
+	fn get_issue_details(&self) -> Option<PullRequestDetails> {
+		parse_pull_request_details_from_url(&self.html_url)
 	}
 }
 
 #[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
 #[serde(rename_all = "snake_case")]
-pub enum WorkflowJobConclusion {
+pub enum GithubWorkflowJobConclusion {
 	#[serde(other)]
 	Unknown,
 }
 #[derive(PartialEq, Deserialize)]
-pub struct WorkflowJob {
+pub struct GithubWorkflowJob {
 	pub head_sha: String,
-	pub conclusion: Option<WorkflowJobConclusion>,
+	pub conclusion: Option<GithubWorkflowJobConclusion>,
 }
 
 #[derive(PartialEq, Deserialize)]
 #[serde(untagged)]
-pub enum Payload {
+pub enum GithubWebhookPayload {
 	IssueComment {
-		action: IssueCommentAction,
-		issue: WebhookIssueComment,
-		comment: Comment,
+		action: GithubIssueCommentAction,
+		issue: GithubWebhookIssueComment,
+		comment: GithubComment,
 	},
 	CommitStatus {
 		// FIXME: This payload also has a field `repository` for the repository where the status
@@ -292,13 +238,13 @@ pub enum Payload {
 		// there's no way to disambiguate between two different PRs in two different repositories with
 		// the same head SHA.
 		sha: String,
-		state: StatusState,
+		state: GithubCommitStatusState,
 	},
 	CheckRun {
-		check_run: CheckRun,
+		check_run: GithubCheckRun,
 	},
-	WorkflowJob {
-		workflow_job: WorkflowJob,
+	GithubWorkflowJob {
+		workflow_job: GithubWorkflowJob,
 	},
 }
 
@@ -311,7 +257,7 @@ struct DetectUserCommentPullRequestPullRequest {
 struct DetectUserCommentPullRequestRepository {
 	pub name: Option<String>,
 	pub full_name: Option<String>,
-	pub owner: Option<User>,
+	pub owner: Option<GithubUser>,
 }
 
 #[derive(Deserialize)]
@@ -327,17 +273,17 @@ struct DetectUserCommentPullRequestComment {
 
 #[derive(Deserialize)]
 pub struct DetectUserCommentPullRequest {
-	action: IssueCommentAction,
+	action: GithubIssueCommentAction,
 	issue: Option<DetectUserCommentPullRequestIssue>,
 	repository: Option<DetectUserCommentPullRequestRepository>,
-	sender: Option<User>,
+	sender: Option<GithubUser>,
 	comment: Option<DetectUserCommentPullRequestComment>,
 }
 
-impl HasIssueDetails for DetectUserCommentPullRequest {
-	fn get_issue_details(&self) -> Option<IssueDetails> {
+impl HasPullRequestDetails for DetectUserCommentPullRequest {
+	fn get_issue_details(&self) -> Option<PullRequestDetails> {
 		if let DetectUserCommentPullRequest {
-			action: IssueCommentAction::Created,
+			action: GithubIssueCommentAction::Created,
 			issue:
 				Some(DetectUserCommentPullRequestIssue {
 					number,
@@ -350,8 +296,8 @@ impl HasIssueDetails for DetectUserCommentPullRequest {
 		} = self
 		{
 			match self.sender {
-				Some(User {
-					type_field: Some(UserType::Bot),
+				Some(GithubUser {
+					type_field: Some(GithubUserType::Bot),
 					..
 				}) => None,
 				_ => {
@@ -359,11 +305,15 @@ impl HasIssueDetails for DetectUserCommentPullRequest {
 
 					if let Some(DetectUserCommentPullRequestRepository {
 						name: Some(name),
-						owner: Some(User { login, .. }),
+						owner: Some(GithubUser { login, .. }),
 						..
 					}) = repository
 					{
-						Some((login.to_owned(), name.to_owned(), *number))
+						Some(PullRequestDetails {
+							owner: login.into(),
+							repo: name.into(),
+							number: *number,
+						})
 					} else {
 						None
 					}
@@ -373,8 +323,13 @@ impl HasIssueDetails for DetectUserCommentPullRequest {
 							..
 						}) = repository
 						{
-							parse_repository_full_name(full_name)
-								.map(|(owner, name)| (owner, name, *number))
+							parse_repository_full_name(full_name).map(
+								|(owner, repo)| PullRequestDetails {
+									owner,
+									repo,
+									number: *number,
+								},
+							)
 						} else {
 							None
 						}
@@ -384,8 +339,7 @@ impl HasIssueDetails for DetectUserCommentPullRequest {
 							html_url: Some(html_url),
 						} = pr
 						{
-							parse_issue_details_from_pr_html_url(html_url)
-								.map(|(owner, name, _)| (owner, name, *number))
+							parse_pull_request_details_from_url(html_url)
 						} else {
 							None
 						}
@@ -398,9 +352,9 @@ impl HasIssueDetails for DetectUserCommentPullRequest {
 	}
 }
 
-pub fn parse_issue_details_from_pr_html_url(
+fn parse_pull_request_details_from_url(
 	pr_html_url: &str,
-) -> Option<IssueDetails> {
+) -> Option<PullRequestDetails> {
 	let re = Regex::new(PR_HTML_URL_REGEX!()).unwrap();
 	let matches = re.captures(pr_html_url)?;
 	let owner = matches.name("owner")?.as_str().to_owned();
@@ -411,10 +365,15 @@ pub fn parse_issue_details_from_pr_html_url(
 		.to_owned()
 		.parse::<i64>()
 		.ok()?;
-	Some((owner, repo, number))
+	Some(PullRequestDetails {
+		owner,
+		repo,
+		number,
+	})
 }
 
-pub fn parse_repository_full_name(full_name: &str) -> Option<(String, String)> {
+/// full_name is org/repo
+fn parse_repository_full_name(full_name: &str) -> Option<(String, String)> {
 	let parts: Vec<&str> = full_name.split('/').collect();
 	parts
 		.get(0)
@@ -429,3 +388,5 @@ pub fn parse_repository_full_name(full_name: &str) -> Option<(String, String)> {
 pub fn owner_from_html_url(url: &str) -> Option<&str> {
 	url.split('/').nth(3)
 }
+
+pub use client::*;
